backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
psycopg2-binary